Python/requirements.txt
requests
Pillow
numpy
torch
torchvision
ultralytics
pyinstaller
